Session 2 - 3g. Int Comparisons: Add integer overloads for number comparisons
Introduce integer-specific comparison logic in G_NumberConditionComparer: add CompareNumberCondition(int, ...) and integer overloads for TestEqual, TestGreater, TestLesser, TestGreaterOrEqual and TestLesserOrEqual (with integer-aware adjustments).
Update G_IntState to call the int overload. Expand and rename unit tests in ConditionTests.cs to cover numerous int comparison scenarios and adjust some test names for clarity.

+++ Assets/Scripts/GOAP/States/G_IntState.cs
@@ -112,7 +112,7 @@ namespace GOAP
         {
             bool result = false;
 
-            result = G_NumberConditionComparer.CompareNumberCondition((float)precondition.ExpectedValue, precondition.Comparison, (float)effect.ExpectedValue, effect.Comparison);
+            result = G_NumberConditionComparer.CompareNumberCondition((int)precondition.ExpectedValue, precondition.Comparison, (int)effect.ExpectedValue, effect.Comparison);
 
             return result;
         }
